requirements.txt
asgiref==3.8.1
boto3==1.34.136
botocore==1.34.136
Django==5.0.6
django-storages==1.14.3
jmespath==1.0.1
pillow==10.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
